Added a column to the Crossover Metrics table produced on STDERR at the end to score the alignment of the segments.
Segments that the user supplied that are supposed to align with one another are scored in pairs by counting the number of alignment positions where a segment of one sequence (as defined by the segments file submitted) aligned with its corresponding segment. Positions with gap charcters are not counted. Whether or not sequences are identical is not taken into consideration.
\nFiles checked in: src/codonHomologizer.pl