Fixed a bug in getExe that prevented -y from working.
Actually, the bug is in the File::Which module and I coded a work-around. I had submitted this some time back to the which developer, but apparently I have some vestigial code lying around that doesn't have my work-around. I also renamed a variable to make its purpose clearer. I formatted the alignment of some code. I also added a note to the verbose output of the weight matrix to indicate that the generated matrix shown is not used if -w was supplied.

Files checked in: src/codonHomologizer.pl