Added some better explanations of what the code is doing in the comments and corrected a minor bug that caused the possibility of not using the best pair of codons as far as minimum usage is concerned.
The best codon pair was saved when comparing the min usage value of the pair to other pairs that had matches in the same places instead of all pairs. The codons swere already in order of descending usage value so the effect of the fix should be small.

Files checked in: src/codonHomologizer.pl