Finished initial debugging. This version works using the test data and either identity segment sizes 5 and 11 or just 11.
This involved major and extensive fixes to logic, but no restructuring was necessary. The overall design was good. A couple notes: I vastly simplified the recoding of the sequences and I apparently wasn't replacing the saved sequence when it was source because I didn't think about the fact that I could be getting sequence from a different alignment other than the source it was supposed to come from. I had some coordinates wrong when determining the closest segment to the left and right. I was getting the query segment coordinate back. I had also been finding the midpoint between the left-most segment and the beginning of the sequence, and vice-versa for the end of the sequence. Those were the latest issues I fixed that were major. I also edited the code that was selecting merge boundaries without regard for codon boundaries.

Files checked in: src/stretchMixer.pl