Resolved 1 issue relating to duplicate dividers between closest segments.
When the first 2 sequence's solution identity segments each calculate a divider differently and copy their divider to their common partner... if that partner is processed after them, it gets stuch with duplicate imprecise dividers because the second divider should have checked whether its (common) partner had a divider in the region where it wanted to create a divider. If it had a divider, it should have converted and used it. I implemented this partner-check for dividers and it appears to work.

Files checked in: src/stretchMixer.pl