Added a check for source divider interruption in dividerExists.
I had to add a new key/value pair in the divider map structure to keep track of "HARD_STOPs" - the place where, when a source divider was created, the source should end based on the identity segments that were used to create it. This check only works for interruptions of existing/already-added dividers. It does not check whether the divider being added goes through an existing RECODE divider.

Cleaned out some old code.

Added a return to getClosestLeft/RightDivider when the range to search is a single position. Even if a divider exists, it would return that divider position anyway, since it's the only option. So there was no point in checking for existing dividers there.

Removed a check on divider type in createRightDividers as a condition before creating a divider. A right divider should only be created if a divider did not exist there already. That was basically the test code that I had previously added, which I commented out after changing the logic.

These changes eliminated 4 "multiple divider" warnings and 1 sequence validation/repair warning.

Files checked in: src/stretchMixer.pl