Multiple bug fixes and catching of divider map issues, mostly relating to problems encountered when copying divider changes from a source divider in one sequence's map to the map of the partner for that identity segment.
I silenced issues that I determined to be innocuous. I fixed some bugs in divider editing and in the last reduction step of the map.

There were 2 warnings that could be significant, but their effect is minimal, given the validation and integrity checks. It only affects non-identity sequences in small portions of sequence in rare instances where the partner divider could not be found because it wasn't in the expected location. The reason it wasn't in the expected location turned out to be a case where either there were gap issues in divider conversion or when a segment was shrunk due to overlap and its partner had a conflict and needed to propagate the change, but couldn't find it due to the shrinking.

It's possible that there could be other related issues, but I'm calling them minor - & not worth the effort.

There are some warnings that will only be issued now when in debug mode.

Files checked in: src/stretchMixer.pl