Fixed a bug that was causing multiple alternate codons to be created in the same position of the same sequence.
There were multiple components to the fix. Alternate codons that were the same as existing codons were not being prevented from being added again simply because they were in agreement/the same. The check of the alternate codon against the interior portion of an identity segment was using the wrong start coordinate. The entire alternate codon has to match the body of a segment. There was a loop control ('next') when the pair ID was the same - which should have only applied to segment records, not alternate codon records.

Files checked in: src/stretchMixer.pl